backend/requirements.txt
fastapi
uvicorn
langchain
langchain-nvidia-ai-endpoints
aiofiles
python-multipart
openai
python-dotenv
PyPDF2
langchain-community
faiss-cpu
sentence-transformers